feat: dynamic settings (#930)
* feat: dynamic settings

* fix: scoping of the matrix

* fix: styling

* refactor: remove initial script

* fix: prevent fatal error when insurance percentage is not an integer

* feat: add priority delivery settings

* refactor: remove unused field from settings

* refactor: rearrange some code

* fix: add french translation

* refactor: more logic names

* docs: remove useless comments

* refactor: remove useless return

* refactor: use json wrapper from magento

+++ src/Model/Source/DefaultOptions.php
@@ -161,7 +161,7 @@ class DefaultOptions
     {
         $total                = $this->quote->getGrandTotal();
         $settings             = $this->config->getCarrierConfig($carrierName, 'default_options');
-        $totalAfterPercentage = $total * (($settings[self::INSURANCE_PERCENTAGE] ?? 0) / 100);
+        $totalAfterPercentage = $total * ((int) ($settings[self::INSURANCE_PERCENTAGE] ?? 0) / 100);
 
         if (! isset($settings[$priceKey])
             || (int) $settings[$priceKey] === 0
